+++ packages/terra-doc-template/tests/wdio/placeholder-spec.js
@@ -1,17 +1,13 @@
-describe('Placeholder', () => {
-  before(() => browser.setViewportSize(Terra.viewports('medium')[0]));
-
+Terra.describeViewports('Placeholder', ['medium'], () => {
   describe('Default Placeholder', () => {
     before(() => browser.url('/#/raw/tests/terra-doc-template/doc-template/default-placeholder'));
 
-    Terra.it.matchesScreenshot();
-    Terra.it.isAccessible();
+    Terra.it.validatesElement();
   });
 
   describe('Light Placeholder', () => {
     before(() => browser.url('/#/raw/tests/terra-doc-template/doc-template/light-placeholder'));
 
-    Terra.it.matchesScreenshot();
-    Terra.it.isAccessible();
+    Terra.it.validatesElement();
   });
 });