+++ packages/terra-i18n/tests/wdio/i18n-spec.js
@@ -1,8 +1,7 @@
 // Add Portuguese-Guinea-Bissau, Zulu & Zulu-South African locales as test locales (supported by intl)
 const testLocales = ['en', 'en-GB', 'pt', 'pt-GW', 'zu', 'zu-ZA'];
 
-describe('I18n', () => {
-  before(() => browser.setViewportSize(Terra.viewports('medium')[0]));
+Terra.describeViewports('I18n', ['medium'], () => {
   describe('Initial Load', () => {
     before(() => {
       browser.url('/#/raw/tests/terra-i-18-n/i-18-n/default-i-18-n');