+++ packages/terra-demographics-banner/CHANGELOG.md
@@ -6,6 +6,9 @@ Unreleased
 ### Updated
 * updated jest snapshots
 
+### Changed
+* Update tests to use `Terra.describeViewports` for setting viewports
+
 3.15.0 - (July 9, 2019)
 ------------------
 ### Changed